+++ Dockerfile
@@ -1,0 +1,4 @@
+FROM ubuntu:latest
+LABEL authors="anyel"
+
+ENTRYPOINT ["top", "-b"]
